ActiveBPELDeployer: getArchiveLocation -> BPR on demand (basic API)
I've been looking for a bit of code which always gets called before
accessing the archive, and this seems to be it. I've touched up all
the calls inside ActiveBPELDeployer so it always calls this one
instead of directly accessing the field.

The basic idea is to allow several levels of control:
- by only setting BPRFile, we use a pre-existing .bpr file as is,
  possibly decorating it a bit with the ActiveBPELDeployment later
  on.
- by only setting BPELFile, a new .bpr is created with the same name
  as the .bpel file, except for the extension. We obtain all the
  information we need starting the .bpel and going through its WSDL
  imports.
- by setting both, the .bpr will always be created from scratch,
  but in the specified path.