IBPELDeployer: getArchiveLocation now takes the PUT
If we're going to implement deployment archive generation for
ActiveBPEL, we'll need to know the port BPELUnit is listening on to.
This information is stored in the PUT object, which is easily
accesible from any of the places where getArchiveLocation is called.

This commit changes the IBPELDeployer interface so getArchiveLocation
takes the PUT instead of the test suite's base path.